backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
python-multipart==0.0.20
openai==1.55.3
pydantic==2.10.3
python-dotenv==1.0.1
fpdf2==2.8.2
aiofiles==24.1.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10
